requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
firebase-admin==6.5.0
psycopg[binary]==3.2.1
langgraph-checkpoint-postgres
